requirements.txt
Flask==2.2.2
requests==2.31.0
Werkzeug==2.2.3
python-dotenv==1.0.1